Fixed bugs: Able to place and break in codespace while not in dev mode Locate message not having bottom line Added event support for: Chat PackLoad Right Click Left Click Click Player Click Entity Place Block Break Block Swap Hands Change Slot Tame

+++ src/main/java/hypersquare/hypersquare/listeners/DevEvents.java
@@ -10,6 +10,7 @@ import hypersquare.hypersquare.plot.CodeBlockManagement;
 import hypersquare.hypersquare.plot.LoadCodeTemplate;
 import hypersquare.hypersquare.plot.RestrictMovement;
 import hypersquare.hypersquare.utils.Utilities;
+import net.minecraft.world.level.block.StainedGlassBlock;
 import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
@@ -36,35 +37,35 @@ public class DevEvents implements Listener {
 
     @EventHandler
     public void onPlayerPlaceBlock(BlockPlaceEvent event) {
-        String plotType = event.getBlock().getWorld().getPersistentDataContainer().get(new NamespacedKey(Hypersquare.getPlugin(Hypersquare.class),"plotType"), PersistentDataType.STRING);
+        String plotType = event.getBlock().getWorld().getPersistentDataContainer().get(new NamespacedKey(Hypersquare.getPlugin(Hypersquare.class), "plotType"), PersistentDataType.STRING);
         RestrictMovement.commonVars(event.getBlockPlaced().getLocation());
         boolean go = true;
-        switch (plotType){
+        switch (plotType) {
 
-            case "Basic" : {
-                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(),RestrictMovement.commonStart,RestrictMovement.basic)) {
+            case "Basic": {
+                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(), RestrictMovement.commonStart, RestrictMovement.basic)) {
                     go = false;
                 }
             }
-            case "Large" : {
-                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(),RestrictMovement.commonStart,RestrictMovement.large)) {
+            case "Large": {
+                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(), RestrictMovement.commonStart, RestrictMovement.large)) {
                     go = false;
 
                 }
 
             }
-            case "Massive" : {
-                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(),RestrictMovement.commonStart,RestrictMovement.massive)) {
+            case "Massive": {
+                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(), RestrictMovement.commonStart, RestrictMovement.massive)) {
                     go = false;
                 }
             }
-            case "Huge" : {
-                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(),RestrictMovement.commonStart,RestrictMovement.huge)) {
+            case "Huge": {
+                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(), RestrictMovement.commonStart, RestrictMovement.huge)) {
                     go = false;
                 }
             }
-            case "Gigantic" : {
-                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(),RestrictMovement.commonStart,RestrictMovement.massive)) {
+            case "Gigantic": {
+                if (!Utilities.locationWithin(event.getBlockPlaced().getLocation(), RestrictMovement.commonStart, RestrictMovement.massive)) {
                     go = false;
                 }
             }
@@ -72,12 +73,14 @@ public class DevEvents implements Listener {
         if (!go)
             event.setCancelled(true);
         if (!Hypersquare.mode.get(event.getPlayer()).equals("coding")) {
+            if (event.getBlock().getLocation().getBlockX() < 0){
+                event.setCancelled(true);
+            }
             return;
         }
-        if (event.getBlock().getLocation().getX() > -0.5){
+        if (event.getBlock().getLocation().getX() > -0.5) {
             return;
         }
-
 
 
         event.setCancelled(true);
@@ -86,12 +89,11 @@ public class DevEvents implements Listener {
         if (event.getBlock().getLocation().getY() < 0 || event.getBlockAgainst().getLocation().getY() < 0 && event.getBlockAgainst().getType() == Material.STONE) {
             return;
         }
-        if (event.getBlock().getLocation().add(1,0,0).getX() > 0){
+        if (event.getBlock().getLocation().add(1, 0, 0).getX() > 0) {
             processPlace(event);
         } else {
-            if (event.getBlock().getLocation().add(1,0,0).getBlock().getType() != Material.AIR && event.getBlockAgainst().getType() != Material.STONE)
-            {
-                Utilities.sendError(event.getPlayer(),"Invalid block placement");
+            if (event.getBlock().getLocation().add(1, 0, 0).getBlock().getType() != Material.AIR && event.getBlockAgainst().getType() != Material.STONE) {
+                Utilities.sendError(event.getPlayer(), "Invalid block placement");
             } else {
                 processPlace(event);
             }
@@ -99,12 +101,11 @@ public class DevEvents implements Listener {
 
     }
 
-    public void processPlace(BlockPlaceEvent event){
-        if (event.getBlock().getLocation().add(-1,0,0).getX() < 0) {
+    public void processPlace(BlockPlaceEvent event) {
+        if (event.getBlock().getLocation().add(-1, 0, 0).getX() < 0) {
             if (ItemManager.getItemID(event.getItemInHand()).startsWith("dev.")) {
 
                 Location location = event.getBlock().getLocation();
-
                 Location againstLocation = event.getBlockAgainst().getLocation();
                 int size = event.getItemInHand().getItemMeta().getPersistentDataContainer()
                         .getOrDefault(new NamespacedKey(plugin, "brackets"), PersistentDataType.STRING, "false").equals("true") ? 4 : 2;
@@ -120,67 +121,73 @@ public class DevEvents implements Listener {
                     Utilities.sendError(event.getPlayer(), "Your code must start with an Event, Function, Process or Entity event.");
                     return;
                 }
+                if (location.clone().add(0, 0, 2).getBlock().getType() == Material.PISTON || location.clone().add(0, 0, 2).getBlock().getType() == Material.STICKY_PISTON) {
+                    CodeBlockManagement.moveCodeLine(location.clone().add(0, 0, 2), 1);
+                }
+
 
                 if (location.getY() >= 0) {
+                    if (location.clone().getBlockY() % 6 == 0) {
+                        Location signLocation = location.clone().add(-1, 0, 0);
+                        String signText = ChatColor.stripColor(event.getItemInHand().getItemMeta().getDisplayName().toUpperCase());
+
+                        signLocation.getBlock().setType(Material.OAK_WALL_SIGN);
+                        BlockData blockData = signLocation.getBlock().getBlockData();
+                        ((Directional) blockData).setFacing(BlockFace.WEST);
+                        Sign sign = (Sign) signLocation.getBlock().getState();
+                        sign.setEditable(true);
+                        sign.getSide(Side.FRONT).setLine(0, signText);
+                        sign.update();
+                        signLocation.getBlock().setBlockData(blockData);
 
 
-                    Location signLocation = location.clone().add(-1, 0, 0);
-                    String signText = ChatColor.stripColor(event.getItemInHand().getItemMeta().getDisplayName().toUpperCase());
+                        new BukkitRunnable() {
+                            @Override
+                            public void run() {
+                                Location codeblockLocation = location.clone();
+                                codeblockLocation.getBlock().setType(event.getItemInHand().getType());
 
-                    signLocation.getBlock().setType(Material.OAK_WALL_SIGN);
-                    BlockData blockData = signLocation.getBlock().getBlockData();
-                    ((Directional) blockData).setFacing(BlockFace.WEST);
-                    Sign sign = (Sign) signLocation.getBlock().getState();
-                    sign.setEditable(true);
-                    sign.getSide(Side.FRONT).setLine(0, signText);
-                    sign.update();
-                    signLocation.getBlock().setBlockData(blockData);
+                                Location stoneLocation = location.clone().add(0, 0, 1);
+                                Location closeBracketLocation = location.clone().add(0, 0, 3);
+                                Location chestLocation = location.add(0, 1, 0);
+                                Location openBracketLocation = location.clone().add(0, -1, 1);
 
+                                if (event.getItemInHand().getItemMeta().getPersistentDataContainer()
+                                        .getOrDefault(new NamespacedKey(plugin, "brackets"), PersistentDataType.STRING, "false").equals("true")) {
+                                    // Open Bracket
+                                    openBracketLocation.getBlock().setType(Material.PISTON);
+                                    BlockData pistonData = openBracketLocation.getBlock().getBlockData();
+                                    ((Directional) pistonData).setFacing(BlockFace.SOUTH);
+                                    openBracketLocation.getBlock().setBlockData(pistonData);
 
+                                    // Close bracket
+                                    closeBracketLocation.getBlock().setType(Material.PISTON);
+                                    pistonData = closeBracketLocation.getBlock().getBlockData();
+                                    ((Directional) pistonData).setFacing(BlockFace.NORTH);
+                                    closeBracketLocation.getBlock().setBlockData(pistonData);
+                                } else {
+                                    // Stone Bracket
+                                    stoneLocation.getBlock().setType(Material.STONE);
+                                }
 
-                    new BukkitRunnable() {
-                        @Override
-                        public void run() {
-                            Location codeblockLocation = location.clone();
-                            codeblockLocation.getBlock().setType(event.getItemInHand().getType());
-
-                            Location stoneLocation = location.clone().add(0, 0, 1);
-                            Location closeBracketLocation = location.clone().add(0, 0, 3);
-                            Location chestLocation = location.add(0, 1, 0);
-                            Location openBracketLocation = location.clone().add(0, -1, 1);
-
-                            if (event.getItemInHand().getItemMeta().getPersistentDataContainer()
-                                    .getOrDefault(new NamespacedKey(plugin, "brackets"), PersistentDataType.STRING, "false").equals("true")) {
-                                // Open Bracket
-                                openBracketLocation.getBlock().setType(Material.PISTON);
-                                BlockData pistonData = openBracketLocation.getBlock().getBlockData();
-                                ((Directional) pistonData).setFacing(BlockFace.SOUTH);
-                                openBracketLocation.getBlock().setBlockData(pistonData);
-
-                                // Close bracket
-                                closeBracketLocation.getBlock().setType(Material.PISTON);
-                                pistonData = closeBracketLocation.getBlock().getBlockData();
-                                ((Directional) pistonData).setFacing(BlockFace.NORTH);
-                                closeBracketLocation.getBlock().setBlockData(pistonData);
-                            } else {
-                                // Stone Bracket
-                                stoneLocation.getBlock().setType(Material.STONE);
+                                if (event.getItemInHand().getItemMeta().getPersistentDataContainer()
+                                        .getOrDefault(new NamespacedKey(plugin, "chest"), PersistentDataType.STRING, "false").equals("true")) {
+                                    chestLocation.getBlock().setType(Material.CHEST);
+                                    BlockData chestBlockData = chestLocation.getBlock().getBlockData();
+                                    ((Directional) chestBlockData).setFacing(BlockFace.WEST);
+                                    chestLocation.getBlock().setBlockData(chestBlockData);
+                                }
                             }
-
-                            if (event.getItemInHand().getItemMeta().getPersistentDataContainer()
-                                    .getOrDefault(new NamespacedKey(plugin, "chest"), PersistentDataType.STRING, "false").equals("true")) {
-                                chestLocation.getBlock().setType(Material.CHEST);
-                                BlockData chestBlockData = chestLocation.getBlock().getBlockData();
-                                ((Directional) chestBlockData).setFacing(BlockFace.WEST);
-                                chestLocation.getBlock().setBlockData(chestBlockData);
-                            }
-                        }
-                    }.runTaskLater(plugin, 1);
+                        }.runTaskLater(plugin, 1);
+                    }
 
                 }
+            } else {
+                Utilities.sendError(event.getPlayer(), "Invalid block placement.");
             }
         }
     }
+
 
     public static void placeCodeTemplate(BlockPlaceEvent event, Plugin plugin) {
         if (event.getItemInHand().getType() == Material.ENDER_CHEST) {
@@ -239,14 +246,17 @@ public class DevEvents implements Listener {
     }
 
 
-
     @EventHandler
     public void onPlayerBreakBlock(BlockBreakEvent event) {
+
         if (!Hypersquare.mode.get(event.getPlayer()).equals("coding")) {
+            if (event.getBlock().getLocation().getBlockX() < 0){
+                event.setCancelled(true);
+            }
             return;
         }
 
-        if (event.getBlock().getLocation().getX() > -0.5){
+        if (event.getBlock().getLocation().getX() > -0.5) {
             return;
         }
 
@@ -256,9 +266,10 @@ public class DevEvents implements Listener {
         if (blockLoc.getY() < 0) {
             return;
         }
+        if (event.getPlayer().getInventory().getItemInMainHand().getType() != Material.AIR) {
             if (Objects.equals(ItemManager.getItemID(event.getPlayer().getInventory().getItemInMainHand()), "dev.glitch")) {
-                if (event.getBlock().getType() == Material.DIAMOND_BLOCK){
-                    Database.removeEventByKey(Utilities.getPlotID(event.getBlock().getWorld()),Utilities.LocationToString(event.getBlock().getLocation()));
+                if (event.getBlock().getType() == Material.DIAMOND_BLOCK) {
+                    Database.removeEventByKey(Utilities.getPlotID(event.getBlock().getWorld()), Utilities.LocationToString(event.getBlock().getLocation()));
                 }
                 event.setCancelled(false);
             } else {
@@ -267,8 +278,8 @@ public class DevEvents implements Listener {
                 }
                 Block block = blockLoc.getBlock();
                 Block signBlock = blockLoc.clone().add(-1, 0, 0).getBlock();
-                if (block.getType() == Material.DIAMOND_BLOCK){
-                    Database.removeEventByKey(Utilities.getPlotID(event.getBlock().getWorld()),Utilities.LocationToString(block.getLocation()));
+                if (block.getType() == Material.DIAMOND_BLOCK) {
+                    Database.removeEventByKey(Utilities.getPlotID(event.getBlock().getWorld()), Utilities.LocationToString(block.getLocation()));
                 }
 
                 if (signBlock.getType() == Material.OAK_WALL_SIGN) {
@@ -279,21 +290,52 @@ public class DevEvents implements Listener {
                     Location bracketLoc = CodeBlockManagement.findCorrespBracket(blockLoc.clone());
                     signBlock.setType(Material.AIR);
                     block.setType(Material.AIR);
+                    if (stoneLoc.getBlock().getType() == Material.PISTON || stoneLoc.getBlock().getType() == Material.STICKY_PISTON) {
+                        bracketLoc.getBlock().setType(Material.AIR);
+                    }
                     stoneLoc.getBlock().setType(Material.AIR);
                     chestLoc.getBlock().setType(Material.AIR);
 
                     if (bracketLoc != null) {
-                        bracketLoc.getBlock().setType(Material.AIR);
-                        CodeBlockManagement.moveCodeLine(bracketLoc.clone().add(0, 0, 1), -1);
                         CodeBlockManagement.moveCodeLine(blockLoc.clone().add(0, 0, 2), -2);
                     } else {
                         CodeBlockManagement.moveCodeLine(blockLoc.clone().add(0, 0, 2), -2);
                     }
                 }
+            }
+        } else {
+            if (blockLoc.getBlock().getType() == Material.OAK_WALL_SIGN) {
+                blockLoc.add(1, 0, 0);
+            }
+            Block block = blockLoc.getBlock();
+            Block signBlock = blockLoc.clone().add(-1, 0, 0).getBlock();
+            if (block.getType() == Material.DIAMOND_BLOCK) {
+                Database.removeEventByKey(Utilities.getPlotID(event.getBlock().getWorld()), Utilities.LocationToString(block.getLocation()));
+            }
 
+            if (signBlock.getType() == Material.OAK_WALL_SIGN) {
+                Location signLoc = signBlock.getLocation();
+                Location chestLoc = blockLoc.clone().add(0, 1, 0);
+                Location stoneLoc = blockLoc.clone().add(0, 0, 1);
 
+                Location bracketLoc = CodeBlockManagement.findCorrespBracket(blockLoc.clone());
+                signBlock.setType(Material.AIR);
+                block.setType(Material.AIR);
+                if (stoneLoc.getBlock().getType() == Material.PISTON || stoneLoc.getBlock().getType() == Material.STICKY_PISTON) {
+                    bracketLoc.getBlock().setType(Material.AIR);
+                }
+                stoneLoc.getBlock().setType(Material.AIR);
+                chestLoc.getBlock().setType(Material.AIR);
 
+                if (bracketLoc != null) {
+                    CodeBlockManagement.moveCodeLine(blockLoc.clone().add(0, 0, 2), -2);
+                } else {
+                    CodeBlockManagement.moveCodeLine(blockLoc.clone().add(0, 0, 2), -2);
+                }
+            }
         }
+
+
     }
 
     @EventHandler
